Move a docstring, add a docstring
